+++ inst/shinyapp/info.php
@@ -1,0 +1,5 @@
+<?php
+$descr ='Collect expert elicitation';
+$version ='ver. 0.1.1 del 2020/01/12';
+$group = 'Basic statistical tools and data reporting';
+?>