+++ NEWS.md
@@ -1,4 +1,10 @@
+* Added download for the pooled plot
 * Bugfix: `dplyr::ungroup()`ed output from `retreived =_options()`
+
+# elicitator 0.1.1
+
+* Fixed button order (by row instead of by column)
+* Added pooled plot
 * Changed SQL query in `retrieve_opinons()` to a full direct SQL call
   to reduce computational and data loading.